Make category and collection assign dialog common
