+++ src/components/AssignContainerDialog/index.ts
@@ -1,0 +1,2 @@
+export * from "./AssignContainerDialog";
+export { default } from "./AssignContainerDialog";
